Fixes based on review
