+++ src/V3EmitMk.cpp
@@ -49,8 +49,7 @@ public:
         of.puts("# C11 constructs required?  0/1 (from --threads, "
                 "--trace-threads or use of classes)\n");
         of.puts("VM_C11 = ");
-        of.puts(v3Global.needC11() || v3Global.opt.threads() || v3Global.opt.traceThreads() ? "1"
-                                                                                            : "0");
+        of.puts(v3Global.needC11() || v3Global.opt.threads() ? "1" : "0");
         of.puts("\n");
         of.puts("# Coverage output mode?  0/1 (from --coverage)\n");
         of.puts("VM_COVERAGE = ");
